Refactor the how we handle the response from FireBase. Nothing was changed logically.
