+++ src/PushNotifications.Contracts/SendPushNotificationResult.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace PushNotifications.Contracts
 {
@@ -16,5 +17,11 @@ namespace PushNotifications.Contracts
         public bool HasFailedTokens { get { return FailedTokens?.Count > 0; } }
 
         public static SendTokensResult Success = new SendTokensResult(new List<SubscriptionToken>());
+
+        /// <summary>
+        /// For some reason the request to the 3rd party failed. However we do nothing in this case for now.
+        /// Ideally we want to execute a retry strategy with Hystrix but it MUST NOT be based on this `Failed` status.
+        /// </summary>
+        public static SendTokensResult Failed = new SendTokensResult(new List<SubscriptionToken>());
     }
 }
