+++ src/PushNotifications.Delivery.FireBase/InternalsVisibleToTests.cs
@@ -1,0 +1,1 @@
+﻿[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("PushNotifications.Delivery.FireBase.Tests")]
